UPDATE: UserLogin.java class - Nathan Binu Edappilly
This is not fully functional yet as we need to implement db logic to store all user and admin login details. In place of this for now I set a Username and Password placeholder that is labelled on the code as a comment. We will have to fully implement this db feature to get user login & registration fully working. See if any of you can get Java Swing running to test out what the front end looks like then me and Harman can get to implementing the db and optimising the front end to look better.