AdminDashboard firestore Parking_spaceslink issue - Nathan
The code implementation looks alright from what I can see, however there are some firestore dependency issues and I can't get it to run. Please look through the comments I have added on AdminDashboard and the xml file. The original working copy of the AdminDashboard is still there, it is just commented out, so if any of you don't manage to fix the dependency/logical issue you can uncomment the original code and get it working as previously again.