UserLogin Update
Factory design pattern implemented for user login in AuthOperation which categorizes registration as 2 options, being able to register as one concrete implementation and logging in as another. The original code for User Login is still present, however it is just commented out in case of any last minute errors with UserLogin.