Created PaymentRates class to handle specific client costs
Formulated in a way where the input is taking from bookingPage class "duration" which is in minutes and uses that input to then convert it into hours(duration in hours), calculation based on client type is then performed and returned. Logic of execution goes into Payment.java that me and Harman will work on. The logic for PaymentRates look accurate, we shall test it and verify the functionality of this class as soon as Payment.java is done being dealt with.